Initial Commit
This site was designed as the final project for my Fundamentals of JavaScript class in the Fall2016. I used HTML5, CSS3, and JavaScript to design this site.
I did all of the graphic work on this site as well. I used Adobe InDesign, Animate and Illustrator to do most of the graphic work for the site.